props/specimens/ducktape/2026-05-20-00/code/props/specimens/ducktape/2026-01-29-00/code/pyproject.toml
# Consolidated Python dependencies for the ducktape monorepo.
# This is the single source of truth for all third-party Python dependencies.
# Individual packages no longer have their own pyproject.toml files.

[project]
name = "ducktape"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
    # Core
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "typer>=0.12.0",
    "typer-di>=0.1.3",
    "click>=8.0",
    "rich>=13.0.0",
    "structlog",
    "platformdirs>=3.0",

    # Web frameworks
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "aiohttp>=3.9.0",
    "requests>=2.32.5",

    # Async
    "anyio>=4.4",
    "aiosqlite>=0.18.0",
    "aiodocker>=0.22.0",
    "asyncpg>=0.29.0",

    # LLM/AI
    "openai>=1.40.0",
    "anthropic",
    "claude-code-sdk>=0.0.21,<0.1",
    "tiktoken>=0.7.0",
    "tenacity>=8.2.0",

    # MCP
    "mcp[cli]>=1.13.1",
    "fastmcp>=2.13",

    # Docker
    "docker>=7.0.0",

    # Data
    "pandas>=2.2.0",
    "PyYAML>=6.0",
    "jinja2>=3.1",
    "pathspec",
    "canonicaljson>=1.6.0",
    "compact-json>=1.4.0",
    "tomlkit",
    "tomli>=2.0.0",
    "tomli-w>=1.0.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.32",
    "alembic>=1.13.0",
    "psycopg[binary]>=3.1",
    "psycopg2-binary>=2.9.6",

    # Git
    "pygit2>=1.18",
    "PyGithub>=1.77",

    # Visualization
    "plotnine>=0.13.0",
    "matplotlib",
    "plotext>=5.2.0",

    # Jupyter
    "ipykernel>=6",
    "jupyter-mcp-server>=0.11.0",
    "jupyter-server",

    # Matrix
    "mautrix>=0.20.0",

    # wt (worktree)
    "colorama>=0.4",
    "pluggy>=1.3",
    "psutil>=5.9",
    "punq>=0.6.2",
    "reaktiv>=0.20",
    "tabulate>=0.9",
    "watchdog>=4.0",

    # gatelet
    "cryptography>=41.0.0",
    "python-multipart>=0.0.9",
    "passlib",
    "fastapi-csrf-protect",
    "fastapi-users",
    "fastapi-users-db-sqlalchemy",
    "fastapi-sessions",
    "aw-client",
    "aw-core",
    "oura",
    "python-metar",
    "pyowm",
    "aqipy",
    "meteostat",
    "geopy",
    "google-api-python-client>=2.0.0",
    "google-auth",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "homeassistant-api",

    # gmail-archiver
    "python-dateutil>=2.8.0",
    "email-reply-parser>=0.5.12",
    "beautifulsoup4>=4.14.3",

    # tana
    "playwright>=1.45",

    # ducktape-llm-common
    "humanize>=4.0.0",
    "pytimeparse>=1.1.8",
    "ruff>=0.5",
    "pre-commit>=3.5",
    "checkov>=3.2",

    # finance
    "absl-py",
    "splitwise",
    "xdg",

    # unidiff
    "unidiff>=0.7.5",

    # props
    "vulture>=2.14",

    # llm/html
    "markdown>=3.5.2",
    "markdownify>=0.11.0",
    "lxml>=5.0.0",

    # habitify
    "python-dotenv",

    # ember
    "matrix-nio>=0.23",
    "jupyter-client>=8.6",
    "ipython>=8.26",
    "kubernetes-asyncio>=26.1",

    # claude_hooks
    "typing-extensions",

    # claude_hooks
    "pproxy>=2.7",
    "pyopenssl>=24.0",
    "pyjks>=20.0",
    "pyjwt>=2.8",
    "supervisor>=4.2",

    # claude_optimizer
    "numpy",

    # git_commit_ai
    "mako>=1.3",

    # gnome-terminal-profile-switcher
    "dbus-python>=1.3.0",
    "PyGObject>=3.40.0",

    # dbus_fast_example
    "dbus-fast",

    # inventree_utils
    "inventree>=0.15",
    "pint>=0.24",
    "tqdm>=4.66",

    # yamllint
    "yamllint>=1.35",
]

[project.optional-dependencies]
# Development and testing dependencies
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "pytest-timeout>=2.3",
    "pytest-xdist>=3.0",
    "pytest-mock>=3.10.0",
    "pytest-env>=1.1.3",
    "pytest-cov>=4.0",
    "pytest-postgresql>=4.1.1",
    "pytest-playwright",
    "pytest-bazel>=0.1.6",
    "testcontainers[postgres]>=4.0.0",
    "asgi-lifespan>=2.1",
    "syrupy>=4.0",
    "PyHamcrest>=2.1",
    "mypy==1.18.2",
    "isort>=5.12.0",
    "invoke",
    "sqlalchemy-stubs>=0.4.0",
]

# Type stubs for mypy
typing = [
    "types-PyYAML",
    "types-requests",
    "types-docker",
    "types-colorama",
    "types-Pygments",
    "types-psycopg2",
    "types-python-dateutil",
    "types-Markdown",
    "types-tabulate",
    "types-filelock",
    "types-toml",
    "types-aiofiles",
    "types-psutil",
    "types-jsonschema",
    "types-unidiff",
    "types-Jinja2",
    "types-sqlalchemy>=1.4.53",
    "asyncpg-stubs",
    "pandas-stubs",
    "sqlalchemy[mypy]>=2.0.32",
]

# Optional: GEPA integration for props
gepa = [
    "gepa>=0.0.22",
    "litellm>=1.0.0",
    "nest-asyncio>=1.5.0",
]

# Optional: Linux desktop notifications
notifications = [
    "dbus-python>=1.2.0",
]

# 2025-12-29: Override rfc3987-syntax with fork that precompiles the Lark parser.
# Replaced N Lark parsers with one precompiled, ~9x smaller import time (0.8s → 0.09s).
# Pending upstream merge: https://github.com/willynilly/rfc3987-syntax/pull/11
[tool.uv]
override-dependencies = ["rfc3987-syntax"]

[tool.uv.sources]
rfc3987-syntax = { git = "https://github.com/agentydragon/rfc3987-syntax", branch = "adgn/one-parser" }

# Tell setuptools this is not a package (just uv config)
# Without this, setuptools discovers all 40+ top-level dirs as packages
[tool.setuptools]
packages = []
